refactor(pds-core): extract preview-shared to fix Sonar duplication gate
PR #103 hit SonarCloud's quality gate at 3.83% new-code duplication
(threshold ≤3%) — driven by a 68-line dup block between
packages/pds-core/src/lib/preview-chooser.ts and preview-consent.ts.
Both routes reconstruct the same authorization-page HTML shell with
identical mechanics (manifest cache, hydration renderer, asset URL
prefix, CSP headers, client-metadata-resolution boilerplate).

Extract a new packages/pds-core/src/lib/preview-shared.ts that owns
the shared mechanics, keeping each route module focused on its
fixture shape and <head> extras. Net -70 LoC across the three files;
the dup block goes away entirely.

Also fold the smaller test-side dup (mockLogger / mockRes in the two
preview test files) into a new __tests__/preview-test-helpers.ts.

Address the one SonarCloud security hotspot too: replace
`Math.random()` with `crypto.randomUUID()` in rate-limit.test.ts's
fake-IP generator. The original wasn't a security boundary but the
swap silences typescript:S2245 without making the test any less
deterministic-in-aggregate.

The remaining two intra-file dup blocks in chooser-enrichment.test.ts
are repeated `it(...)` shapes asserting on
`buildChooserEnrichmentScript()` output; deduping them would hurt
test-as-prose readability and they don't push the percentage back
over the gate.
