+++ packages/pds-core/src/__tests__/preview-consent.test.ts
@@ -4,31 +4,7 @@ import {
   createPreviewConsentHandler,
   renderPreviewIndex,
 } from '../lib/preview-consent.js'
-
-function mockLogger() {
-  return { info: vi.fn(), warn: vi.fn(), debug: vi.fn() }
-}
-
-type CapturedRes = {
-  headers: Record<string, string>
-  body: string | null
-  setHeader: (name: string, value: string) => void
-  send: (body: string) => void
-}
-
-function mockRes(): CapturedRes {
-  const res: CapturedRes = {
-    headers: {},
-    body: null,
-    setHeader(name, value) {
-      this.headers[name] = value
-    },
-    send(body) {
-      this.body = body
-    },
-  }
-  return res
-}
+import { mockLogger, mockRes } from './preview-test-helpers.js'
 
 describe('createPreviewConsentHandler', () => {
   // Snapshot + restore per-test so a mid-test throw cannot leak env state
